requirements.txt
pandas
numpy
matplotlib
seaborn
boto3
python-dotenv
